Fix buffering and exception handling for json in solr
Fixes three bugs with solr.  First, that exceptions being thrown
while JSON responses were generated were being hidden.

Second, makes it so that even if a doc has attributes no in the schema
it doesn't break the results.  I copied the pattern from XMLWriter in
the same package (the pattern is also used later in the file).

Third, it works around a bug with buffering by putting some flushes
strategically in the file.